DLM
Introducing, DLM.
DLM stands for, 'Dynamically Loading Modules'.
Why? Cause cogs weren't working so I made this... it works, it's 'easy' to add to. Just make a .py file. throw it in, and watch it work. This will also move us up to FrogBot v2.1, and I expect this to break and have lots of bugs at first, which I do not look forward to fixing. But it's for the future of FrogBot.